requirements.txt
# Core dependencies
cohere>=5.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Optional: Para mejorar el proyecto
# chromadb>=0.4.0  # Base de datos vectorial
# streamlit>=1.28.0  # Interfaz web
